Move Chained.reduce (prev findChain) into Chained
