+++ src/index.ts
@@ -4,4 +4,4 @@ export * as keypair from "./keypair"
 export * from "./keypair/ed25519"
 export * from "./keypair/rsa"
 export * from "./types"
-export * from "./chain"
+export * from "./chained"
